Migrate API-based settings payloads so the API can now receive a settings payload either from an earlier version, or one without a version at all.
Migrating deployment profiles is covered in issue 5901

To reiterate, this means that the version field is no longer needed
in a settings payload, but it's more efficient to specify it.

I took the version field out of the bats tests, because these
tests should be as simple as possible, but it's left in `rdctl set`,
because we want that to be as efficient as possible. Also, leaving
it in finds the problem where someone is using a version of rdctl
with a newer prefs version than the server, which isn't supported.

Signed-off-by: Eric Promislow <epromislow@suse.com>